:sparkles: feat: crimplementação das exceptions

+++ src/main/java/com/projeto/saude_hub/service/impl/ConsultaServiceImpl.java
@@ -4,10 +4,13 @@ import com.projeto.saude_hub.controller.dto.ConsultaDto;
 import com.projeto.saude_hub.domain.model.consulta.Consulta;
 import com.projeto.saude_hub.domain.model.consulta.StatusConsulta;
 import com.projeto.saude_hub.domain.repository.ConsultaRepository;
+import com.projeto.saude_hub.exceptions.consulta.CamposNulosConsultaException;
+import com.projeto.saude_hub.exceptions.consulta.ConsultaNaoEncontradaException;
 import com.projeto.saude_hub.service.ConsultaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.lang.reflect.Field;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -24,6 +27,10 @@ public class ConsultaServiceImpl implements ConsultaService {
 
     @Override
     public Consulta create(ConsultaDto consultaDTO) {
+        if(hasCamposNulos(consultaDTO)){
+            throw new CamposNulosConsultaException(consultaDTO);
+        }
+
         Consulta consulta = new Consulta(
                 null,
                 consultaDTO.dataConsulta(),
@@ -41,6 +48,11 @@ public class ConsultaServiceImpl implements ConsultaService {
 
     @Override
     public Optional<ConsultaDto> findById(Long id) {
+        Optional<Consulta> consulta = consultaRepository.findById(id);
+
+        if(consulta.isEmpty()){
+            throw new ConsultaNaoEncontradaException(id);
+        }
         return consultaRepository.findById(id).map(ConsultaDto::new);
     }
 
@@ -54,6 +66,7 @@ public class ConsultaServiceImpl implements ConsultaService {
     @Override
     public List<ConsultaDto> findByStatus(StatusConsulta status) {
         List<Consulta> consultas = consultaRepository.findByStatus(status);
+
         return consultas.stream().map(ConsultaDto::new).collect(Collectors.toList());
     }
 
@@ -61,7 +74,7 @@ public class ConsultaServiceImpl implements ConsultaService {
     @Override
     public Optional<ConsultaDto> update(Long id, ConsultaDto consultaDto) {
         Consulta consulta = consultaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Consulta não encontrada"));
+                .orElseThrow(() -> new ConsultaNaoEncontradaException(id));
 
         consulta.setDataConsulta(consultaDto.dataConsulta());
         consulta.setEspecialidade(consultaDto.especialidade());
@@ -77,6 +90,31 @@ public class ConsultaServiceImpl implements ConsultaService {
 
     @Override
     public void delete(Long id) {
+        Optional<Consulta> consulta = consultaRepository.findById(id);
+
+        if(consulta.isEmpty()){
+            throw new ConsultaNaoEncontradaException(id);
+        }
+
         consultaRepository.deleteById(id);
     }
+
+    private boolean hasCamposNulos(ConsultaDto consultaDTO) {
+        for (Field field : consultaDTO.getClass().getDeclaredFields()) {
+            if (!CamposNulosConsultaException.getCamposObrigatorios().contains(field.getName())) {
+                continue;
+            }
+
+            field.setAccessible(true);
+            try {
+                if (field.get(consultaDTO) == null) {
+                    return true;
+                }
+            } catch (IllegalAccessException e) {
+                throw new RuntimeException("Erro ao verificar campos nulos.", e);
+            }
+        }
+        return false;
+    }
+
 }
